feature/ui : Implement Bottom Tab Bar
ADD Bottom Tab Bar Components
ADD MainView Login Conditional View Flow

TODO - ADD Product Detail View

+++ Panda-Saza/Panda-Saza/Models/ShopItem/ItemThumbnail.swift
@@ -14,7 +14,7 @@ struct ItemThumbnail: Hashable, Codable {
     var itemNameTrans: String
     var itemPrice: Int
     
-    var registrationTime: Int
+    var registrationTime: Double
     
     var sellerLoc: String
     var interest: Int
